Rename classes slightly and break the code apart into multiple files
Rename a few things at the same time:

* “Injector” is the object connected to the container ready to be used inside your app’s classes
* “Injection” is the module that is included into your class when the injector is called with the names of the class’s dependencies

This also leaves `Dry::AutoInject` available is a mostly-empty module that we can use to hold our utility method for finding the arity of `super` methods.
